requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
easyocr==1.7.0
filelock==3.12.3
idna==3.4
imageio==2.31.2
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
pyclipper==1.3.0.post4
python-bidi==0.4.2
pytube3==9.6.4
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.8.30
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4